Use m_pSCTHelper->strip((*rdoItr)->identify()) instead of (*rdoItr)->getStrip() in src/SCT_CalibHitmapSvc.cxx and src/SCT_CalibLbSvc.cxx. (SCT_CalibAlgs-00-03-44)
	* Use m_pSCTHelper->strip((*rdoItr)->identify()) instead of
	  (*rdoItr)->getStrip() in src/SCT_CalibHitmapSvc.cxx and
	  src/SCT_CalibLbSvc.cxx.
	  https://its.cern.ch/jira/browse/ATLASSIM-1907
	* tag as SCT_CalibAlgs-00-03-44

2016-03-10 Shaun Roe
	* fix compiler warnings and coverity defects 105663, 105662:
105663 08/02/2016 (Medium) Result is not floating-point :/InnerDetector/InDetCalibAlgs/SCT_CalibAlgs/src/SCTCalib.cxx in function "getNumNoisyStrips"
105662 08/02/2016 (Medium) Result is not floating-point :/InnerDetector/InDetCalibAlgs/SCT_CalibAlgs/src/SCT_CalibHvSvc.cxx in function "fill"
	* tag as SCT_CalibAlgs-00-03-43

2016-02-24 Alberto Gascon
	* Generation of 2D hitmaps as a function of lumiblocks added.
	* Tagged as SCT_CalibAlgs-00-03-42

2016-02-17 Alberto Gascon
	* Generation of 2D hitmaps temporarily removed. Needs checking
	* Tagged as SCT_CalibAlgs-00-03-41

...
(Long ChangeLog diff - truncated)

+++ InnerDetector/InDetCalibAlgs/SCT_CalibAlgs/src/SCT_CalibHvSvc.cxx
@@ -168,7 +168,7 @@ SCT_CalibHvSvc::fill(const bool fromData){
       tq[iq]=m_phvtripQueue[waferhash].front();
       m_phvtripQueue[waferhash].pop();
     }
-    if (tq[m_maxq-2]/m_maxtbins > limit && tq[m_maxq-3]/m_maxtbins > limit && tq[m_maxq-4]/m_maxtbins > limit){
+    if ((double)tq[m_maxq-2]/m_maxtbins > limit && tq[m_maxq-3]/m_maxtbins > limit && tq[m_maxq-4]/m_maxtbins > limit){
         // found a persistent trip/thing
       if (msgLvl(MSG::DEBUG)) msg(MSG::DEBUG) << "Found a potential trip in SCT wafer "<< waferhash<<endreq;
             //if (msgLvl(MSG::DEBUG)) msg(MSG::DEBUG) << "TRIP: "<< waferhash<<" "<<time_stamp <<" "<< Event_number <<" "<< (numhits + tq[m_maxq-1] ) <<endreq;
